moved the api for contextBridge to a module that can be re-used during jest setup. #870.
