+++ workbench/babel.config.js
@@ -1,6 +1,7 @@
 module.exports = {
   presets: [
     '@babel/preset-env',
+    '@babel/preset-react', // needed only for tests
   ],
   plugins: [
     '@babel/plugin-transform-runtime',
